simulate

+++ simulate/simulation.py
@@ -19,9 +19,10 @@ save_path = BASE_DIR + '/../data/simulation/'
 
 filepath = save_path + 'true_net_1000x300.csv'
 
-description = '1000x300'
-nodes_num = 300
-node_dim = 2  # 两个维度
+K=2
+description = '2000x300'
+nodes_num = 300 * K
+node_dim = 2
 time = 10
 dt = 0.01
 if os.path.exists(filepath) and False:
@@ -29,13 +30,11 @@ if os.path.exists(filepath) and False:
     nodes = file_net[['node1_x', 'node1_y']].iloc[range(nodes_num)].values
     val_hidden = file_net[description].values
 else:
-    # 1. generate nodes
+    # 1. generate nodes，这里的点就是用户向量，只不过是二维
     nodes = random.rand(nodes_num, node_dim)
-    print(nodes)
 
     val_hidden = lambda x: 1 - sqrt(sum((x[:node_dim] - x[node_dim:]) ** 2)) / sqrt(2.)
     net_hidden1 = generate_network(nodes, val_hidden)
-exit(0)
 
 # 2. max possible edges counts, gererate those index
 evl = []
@@ -49,7 +48,7 @@ initial = zeros(nodes_num)
 for i in range(0, 50):  # 初始感染前50个节点
     initial[i] = 1
 
-network = network_estimation(time, dt, nodes, val_hidden, trails=2000, band_power=1. / float(node_dim + 1))
+network = network_estimation(time, dt, nodes, val_hidden, trails=2000, band_power=1. / float(node_dim + 1), K=K)
 solutions, time_line = network.simulation(val_hidden, nodes, initial, time, dt, 0, array([[]]), 2, net1=None, net2=None,
                                           true_net=False, hidden_network_fun=val_hidden)
 obs_t = time_line
@@ -63,11 +62,11 @@ true_net['net_hidden'] = network.net2.flatten()
 
 net_hidden = array(true_net['net_hidden']).reshape(nodes_num, nodes_num)
 
-if os.path.exists(filepath):
-    pd.DataFrame(obs_t).to_csv(save_path + 'obst_1000x300_' + description + '_re' + '.csv')  # 这个没用了
-    pd.DataFrame(obs).to_csv(save_path + 'obs_1000x300_' + description + '_re' + '.csv')
+if os.path.exists(filepath) and False:
+    # pd.DataFrame(obs_t).to_csv(save_path + 'obst_1000x300_' + description + '_re' + '.csv')  # 这个没用了
+    pd.DataFrame(obs).to_csv(save_path + 'obs_000x300_' + description + '_re' + '.csv')
     true_net.to_csv(save_path + 'true_net_1000x300_' + description + '_re' + '.csv')
 else:
-    pd.DataFrame(obs_t).to_csv(save_path + 'obst_' + description + '.csv')
+    # pd.DataFrame(obs_t).to_csv(save_path + 'obst_' + description + '.csv')
     pd.DataFrame(obs).to_csv(save_path + 'obs_' + description + '.csv')
     true_net.to_csv(save_path + 'true_net_' + description + '.csv')
